+++ doc/whats-new.rst
@@ -15,9 +15,16 @@ What's New
 
 .. _whats-new.0.9.7:
 
-v0.9.7 (unreleased)
+v0.10.0 (unreleased)
 -------------------
 
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+- Supplying ``coords`` as a dictionary to the ``DataArray`` constructor without
+  also supplying an explicit ``dims`` argument is no longer supported. This
+  behavior was deprecated in version 0.9 but is now an error (:issue:`727`).
+  By `Joe Hamman <https://github.com/jhamman>`_.
 Backward Incompatible Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~